THRIFT-2180: changes TTransport methods to use size_t
size_t is conveniently unsigned and the maximum allowed width
on the target platform. This change makes it possible to
preserve this width across all implementors of the protocol.

+++ lib/cocoa/src/transport/TFramedTransport.m
@@ -73,12 +73,13 @@
     [writeBuffer appendBytes:dummy_header length:HEADER_SIZE];
 }
 
-- (void)write:(const uint8_t *)data offset:(unsigned int)offset length:(unsigned int)length
+- (void) write: (const uint8_t *) data offset: (size_t) offset length: (size_t) length
 {
     [writeBuffer appendBytes:data+offset length:length];
 }
 
-- (int)readAll:(uint8_t *)buf offset:(int)off length:(int)len {
+- (size_t) readAll: (uint8_t *) buf offset: (size_t) off length: (size_t) len
+{
     if (readBuffer == nil) {
         [self readFrame];
     }
